% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bu_map.R
\name{bu_map}
\alias{bu_map}
\title{Displays an interactive map of breeding locations}
\usage{
bu_map(locs)
}
\arguments{
\item{locs}{brapi location object}
}
\description{
Displays an interactive map of breeding locations
}
